Fixes #31; We now have PinkKey, PinkWall-Block, and PinkWall-Phase
Prefab PinkKey will disable all blocks with tag PinkStructure, PinkWall-Block prefab will be baked as a wall in NavMesh, PinkWall-Phase will be baked as air in NavMesh, meaning robots will phase through the Phase-walls.
This requires the Phase layer be not included in the navmesh bake, as well as Block being included.