Level Updates
Hey, I just made the two levels I made easier.
The first one has more space behind the box on the way to the exit from the pink key and the other one has more time react when you pick up the key and the robots start the chase.
The speed of the enemy was changed, of Brad and the size of the key and folder.