Lighting update
I have added working lighting on the 1st level, it is just for reference.
I changed the floor's texture to a basic one for testing purposes.
Also, updated the pink door's texture as it seemed to bug with the lighting/.
Lastly, I moved the doors on the 1st level down to the floor so that enemies don't see you through the walls.